+++ packages/compilers/CHANGELOG.md
@@ -1,3 +1,9 @@
+
+
+## @ethereum-sourcify/compilers@1.1.1 - 2026-04-14
+
+- update dependencies
+
 ## @ethereum-sourcify/compilers@1.1.0 - 2026-03-30
 
 - feat: add Fe language support for smart contract verification (#2692)